// src/app/Models/QuickAdmission.ts
export interface QuickAdmission {
    Id: number;
    UserId: number;
    firstName: string;
    lastName: string;
    email: string;
    fatherName: string;
    gender: string;
    phone: string;
    created: Date;
    dateofBirth: Date;
    doa:Date;
    motherTongue: string;
    rollNo:string;
    standers: string;
    classsection: string;
    classLevel: string;
    admissionNo: string;
    Nationality: string;
    medium: string;
}
 